Added prompts for environment, services, tags and terraform commands.
